refactor: decouple capability repository
