+++ src/apps/server/experiences/providers/experience.provider.ts
@@ -1,5 +1,6 @@
 import { AiResumeRepository } from '@libs/modules/database/repositories/aiResume/aiResume.interface';
 import { AiResumeRepositoryImpl } from '@libs/modules/database/repositories/aiResume/aiResume.repository';
+import { CapabilityRepository } from '@libs/modules/database/repositories/capability/capability.interface';
 import { CapabilityRepositoryImpl } from '@libs/modules/database/repositories/capability/capability.repository';
 import { ExperienceRepository } from '@libs/modules/database/repositories/experience.repository';
 import { ExperienceCapabilityRepository } from '@libs/modules/database/repositories/experienceCapability.repository';
@@ -8,7 +9,10 @@ import { ExperienceInfoRepository } from '@libs/modules/database/repositories/ex
 export const experienceProviders = [
   ExperienceRepository,
   ExperienceInfoRepository,
-  CapabilityRepositoryImpl,
+  {
+    provide: CapabilityRepository,
+    useClass: CapabilityRepositoryImpl,
+  },
   ExperienceCapabilityRepository,
   {
     provide: AiResumeRepository,
