+++ src/libs/modules/database/repositories/capability/capability.interface.ts
@@ -1,0 +1,11 @@
+import { Capability, Prisma } from '@prisma/client';
+
+export interface CapabilityRepository {
+  findMany(condition: { where: Partial<Capability> });
+  findFirst(condition: { where: Partial<Capability> });
+  create(data: Prisma.CapabilityCreateArgs);
+  countExperienceAndCapability(userId: number, isCompleted?: boolean);
+  findByUserId(userId: number): Promise<Capability[]>;
+}
+
+export const CapabilityRepository = Symbol('CapabilityRepository');
